requirements.txt
matplotlib~=3.3.4
numpy~=1.20.3
pillow~=9.3.0
torchvision~=0.14.0
pyyaml~=6.0
scikit-learn~=1.1.3
torchinfo~=1.7.1
geopandas~=0.9.0
pandas~=1.2.4